+++ apps/AppWithWearable/lib/pages/home/page.dart
@@ -4,14 +4,13 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:flutter_foreground_task/flutter_foreground_task.dart';
-import 'package:friend_private/backend/api_requests/api_calls.dart';
+import 'package:friend_private/backend/api_requests/api/other.dart';
 import 'package:friend_private/backend/api_requests/cloud_storage.dart';
 import 'package:friend_private/backend/database/memory.dart';
 import 'package:friend_private/backend/database/memory_provider.dart';
 import 'package:friend_private/backend/mixpanel.dart';
 import 'package:friend_private/backend/preferences.dart';
 import 'package:friend_private/backend/schema/bt_device.dart';
-import 'package:friend_private/backend/storage/message.dart';
 import 'package:friend_private/pages/capture/page.dart';
 import 'package:friend_private/pages/capture/widgets/transcript.dart';
 import 'package:friend_private/pages/chat/page.dart';
@@ -27,20 +26,6 @@ import 'package:friend_private/utils/sentry_log.dart';
 import 'package:gradient_borders/gradient_borders.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:upgrader/upgrader.dart';
-import 'package:workmanager/workmanager.dart';
-
-@pragma('vm:entry-point')
-void callbackDispatcher() {
-  Workmanager().executeTask((task, inputData) async {
-    debugPrint("Native called background task: $task");
-    // createNotification(
-    //   title: 'Here is your action plan for tomorrow',
-    //   body: 'Check out your daily summary to see what you should do tomorrow.',
-    //   notificationId: 5,
-    // );
-    return Future.value(true);
-  });
-}
 
 class HomePageWrapper extends StatefulWidget {
   final dynamic btDevice;
@@ -58,8 +43,6 @@ class _HomePageWrapperState extends State<HomePageWrapper> with WidgetsBindingOb
 
   List<Memory> memories = [];
 
-  // List<Message> messages = [];
-
   FocusNode chatTextFieldFocusNode = FocusNode(canRequestFocus: true);
   FocusNode memoriesTextFieldFocusNode = FocusNode(canRequestFocus: true);
 
@@ -76,12 +59,6 @@ class _HomePageWrapperState extends State<HomePageWrapper> with WidgetsBindingOb
     memories = (await MemoryProvider().getMemoriesOrdered(includeDiscarded: true)).reversed.toList();
     setState(() {});
   }
-
-  // _loadMessages() async {
-  //   var msgs = SharedPreferencesUtil().chatMessages;
-  //   messages = msgs.isEmpty ? [Message(text: 'What would you like to search for?', type: 'ai', id: '1')] : msgs;
-  //   setState(() {});
-  // }
 
   _setupHasSpeakerProfile() async {
     // SharedPreferencesUtil().hasSpeakerProfile = await userHasSpeakerProfile(SharedPreferencesUtil().uid);
@@ -110,29 +87,6 @@ class _HomePageWrapperState extends State<HomePageWrapper> with WidgetsBindingOb
     _initiateMemories();
   }
 
-  _initDailySummaries() {
-    debugPrint('_initDailySummaries');
-    // Workmanager().initialize(callbackDispatcher, isInDebugMode: true);
-    var now = DateTime.now();
-    var target = DateTime(now.year, now.month, now.day, 20, 0);
-    if (now.isAfter(target)) target = target.add(const Duration(days: 1));
-    var delay = target.difference(now);
-    createNotification(
-      title: 'Don\'t forget to wear Friend today',
-      body: 'Wear your friend and capture your memories today.',
-      notificationId: 4,
-      isMorningNotification: true,
-    );
-    // Workmanager().registerPeriodicTask(
-    //   "com.friend-app-with-wearable.ios12.daily-summary",
-    //   "com.friend-app-with-wearable.ios12.daily-summary",
-    //   frequency: const Duration(seconds: 60), // not considered for iOS
-    //   initialDelay: const Duration(seconds: 0),
-    //   constraints: Constraints(networkType: NetworkType.connected),
-    // );
-    // debugPrint('_initDailySummaries registered');
-  }
-
   @override
   void initState() {
     _controller = TabController(length: 3, vsync: this, initialIndex: 1);
@@ -144,7 +98,6 @@ class _HomePageWrapperState extends State<HomePageWrapper> with WidgetsBindingOb
     Upgrader.clearSavedSettings();
 
     _initiateMemories();
-    // _loadMessages();
     _initiatePlugins();
     _setupHasSpeakerProfile();
     _migrationScripts();
@@ -163,7 +116,6 @@ class _HomePageWrapperState extends State<HomePageWrapper> with WidgetsBindingOb
       isDailySummaryNotification: true,
     );
 
-    // _initDailySummaries();
     super.initState();
   }
